pyproject.toml
[project]
name = "data-engineering-zoomcamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.4",
    "google-cloud-storage>=3.9.0",
    "pandas>=3.0.0",
]

[tool.uv.workspace]
members = [
    "03-data-warehouse/extras",
]
